Reorganize to a volume module
