+++ NewHorizons/Handlers/PlanetCreationHandler.cs
@@ -3,6 +3,7 @@ using NewHorizons.Builder.Body;
 using NewHorizons.Builder.General;
 using NewHorizons.Builder.Orbital;
 using NewHorizons.Builder.Props;
+using NewHorizons.Builder.Volumes;
 using NewHorizons.Components;
 using NewHorizons.Components.Orbital;
 using NewHorizons.OtherMods.OWRichPresence;
@@ -606,6 +607,11 @@ namespace NewHorizons.Handlers
             if (body.Config.Props != null)
             {
                 PropBuildManager.Make(go, sector, rb, body.Config, body.Mod);
+            }
+
+            if (body.Config.Volumes != null)
+            {
+                VolumesBuildManager.Make(go, sector, body.Config, body.Mod);
             }
 
             if (body.Config.Funnel != null)
